chore: move "/scripts/docs/" content to "/docs/source/", also re-add sub-directory files
