+++ docs/source/home.js
@@ -1,0 +1,3 @@
+const package = require('./../../package.json')
+
+module.exports = { package: package, title: 'ODM' };
