+++ docs/source/tutorials.js
@@ -1,0 +1,7 @@
+const { mapSubDoc } = require("./utils");
+
+mapSubDoc('tutorials', {
+    title: `Mongoose Tutorials:`,
+    acquit: true,
+    markdown: true
+  }, module.exports);
