+++ docs/source/typescript.js
@@ -1,0 +1,6 @@
+const { mapSubDoc } = require("./utils");
+
+mapSubDoc('typescript', {
+    title: `Mongoose:`,
+    markdown: true
+  }, module.exports);
